Add output routines to allow consistent formatting of memory, strings and bignums.  These have been refactored into their own file, along with their error displays.  The formatting follows the output format used on error, except that bignums of sixty four bits or less are displayed in a more compact one line form.
Added a TEST_note function for producing output without file and line
information.

Update the three tests that call BN_print so they use the new test
infrastructure instead.
